refactor(api): migrate schedules list and patch to Effect programs
Extract listSchedules and patchSchedule into Effect programs with
Schema.TaggedError domain errors. Create/delete routes remain Promise-based.
