+++ apps/api/src/routes/schedules.ts
@@ -1,34 +1,25 @@
 import crypto from "node:crypto";
 
 import { createRoute } from "@hono/zod-openapi";
-import { toUtcDateString } from "@notra/ai/utils/schedule-interval";
 import type { createDb } from "@notra/db/drizzle";
 import {
   contentTriggerLookbackWindows,
   contentTriggers,
-  githubIntegrations,
 } from "@notra/db/schema";
 import {
   createScheduleRequestSchema,
-  scheduleTargetsRepositoryIdsSchema,
   deleteScheduleResponseSchema,
   getSchedulesQuerySchema,
   getSchedulesResponseSchema,
   patchScheduleRequestSchema,
-  scheduleOutputConfigSchema,
   scheduleParamsSchema,
   scheduleResponseSchema,
-  scheduleSourceConfigSchema,
-  scheduleTargetsSchema,
 } from "@notra/schemas/api/schedules";
-import { and, desc, eq, inArray, ne } from "drizzle-orm";
+import { and, eq } from "drizzle-orm";
 // biome-ignore lint/performance/noNamespaceImport: Zod recommended way of importing
 import * as z from "zod";
 
-import type {
-  ScheduleTriggerRow,
-  ScheduleTriggerWithLookbackWindow,
-} from "../types/schedules";
+import { listSchedules, patchSchedule } from "../programs/schedules";
 import { getOrganizationId } from "../utils/auth";
 import { logError } from "../utils/logging";
 import { createOpenApiApp } from "../utils/openapi-app";
@@ -39,185 +30,21 @@ import {
   ORGANIZATION_SCHEDULE_PATH_REGEX,
   ORGANIZATION_SCHEDULES_PATH_REGEX,
 } from "../utils/regex";
+import {
+  DEFAULT_SCHEDULE_NAME,
+  ensureScheduleTargetsExist,
+  hashSchedule,
+  mapQstashError,
+  normalizeSchedule,
+  runScheduleProgram,
+  serializeSchedule,
+} from "../utils/schedules";
 import { deleteQstashScheduleWithRetry } from "../utils/triggers";
 
 export const schedulesRoutes = createOpenApiApp();
 
 type DbClient = ReturnType<typeof createDb>;
 type CreateScheduleBody = z.infer<typeof createScheduleRequestSchema>;
-
-const DEFAULT_SCHEDULE_NAME = "Untitled Schedule";
-
-function normalizeCronConfig(
-  config: CreateScheduleBody["sourceConfig"]["cron"]
-) {
-  const base = {
-    frequency: config.frequency,
-    hour: config.hour,
-    minute: config.minute,
-  } as const;
-
-  if (config.frequency === "weekly") {
-    return {
-      ...base,
-      dayOfWeek: config.dayOfWeek ?? 1,
-    };
-  }
-
-  if (config.frequency === "monthly") {
-    return {
-      ...base,
-      dayOfMonth: config.dayOfMonth ?? 1,
-    };
-  }
-
-  if (config.frequency === "custom") {
-    return {
-      ...base,
-      intervalDays: config.intervalDays,
-      anchorDate: config.anchorDate ?? toUtcDateString(new Date()),
-    };
-  }
-
-  return base;
-}
-
-function normalizeSchedule(input: CreateScheduleBody) {
-  return {
-    ...input,
-    sourceConfig: {
-      cron: normalizeCronConfig(input.sourceConfig.cron),
-    },
-    targets: {
-      repositoryIds: [...input.targets.repositoryIds].sort(),
-    },
-  };
-}
-
-function hashSchedule(input: CreateScheduleBody) {
-  const normalized = normalizeSchedule(input);
-
-  return crypto
-    .createHash("sha256")
-    .update(
-      JSON.stringify({
-        sourceType: normalized.sourceType,
-        sourceConfig: normalized.sourceConfig,
-        targets: normalized.targets,
-        outputType: normalized.outputType,
-        outputConfig: normalized.outputConfig ?? null,
-        lookbackWindow: normalized.lookbackWindow,
-      })
-    )
-    .digest("hex");
-}
-
-async function ensureScheduleTargetsExist(
-  db: DbClient,
-  organizationId: string,
-  repositoryIds: string[],
-  message: string
-) {
-  if (repositoryIds.length === 0) {
-    return null;
-  }
-
-  const integrations = await db.query.githubIntegrations.findMany({
-    where: and(
-      eq(githubIntegrations.organizationId, organizationId),
-      inArray(githubIntegrations.id, repositoryIds)
-    ),
-    columns: { id: true },
-  });
-
-  const existingIds = new Set(
-    integrations.map((integration) => integration.id)
-  );
-  const missingIds = repositoryIds.filter((id) => !existingIds.has(id));
-
-  if (missingIds.length > 0) {
-    return { error: message, missingIntegrationIds: missingIds };
-  }
-
-  return null;
-}
-
-function serializeSchedule(trigger: ScheduleTriggerWithLookbackWindow) {
-  return {
-    id: trigger.id,
-    organizationId: trigger.organizationId,
-    name: trigger.name,
-    sourceType: "cron" as const,
-    sourceConfig: scheduleSourceConfigSchema.parse(trigger.sourceConfig),
-    targets: scheduleTargetsSchema.parse(trigger.targets),
-    outputType: createScheduleRequestSchema.shape.outputType.parse(
-      trigger.outputType
-    ),
-    outputConfig:
-      trigger.outputConfig == null
-        ? null
-        : scheduleOutputConfigSchema.parse(trigger.outputConfig),
-    enabled: trigger.enabled,
-    autoPublish: trigger.autoPublish,
-    createdAt: trigger.createdAt.toISOString(),
-    updatedAt: trigger.updatedAt.toISOString(),
-    lookbackWindow: trigger.lookbackWindow,
-  };
-}
-
-function safeSerializeSchedule(
-  trigger: Parameters<typeof serializeSchedule>[0]
-) {
-  try {
-    return serializeSchedule(trigger);
-  } catch (error) {
-    if (!(error instanceof z.ZodError)) {
-      throw error;
-    }
-
-    logError(`Skipping malformed schedule ${trigger.id}`, error);
-    return null;
-  }
-}
-
-function mapQstashError(error: unknown) {
-  const message = error instanceof Error ? error.message : "Unknown error";
-
-  if (
-    message.includes("invalid destination") ||
-    message.includes("unable to resolve host") ||
-    message.includes("WORKFLOW_BASE_URL is not configured")
-  ) {
-    return { error: "External URL not configured", status: 400 as const };
-  }
-
-  return { error: "Failed to configure schedule", status: 500 as const };
-}
-
-function filterByRepositoryIds<T extends ScheduleTriggerRow>(
-  triggers: T[],
-  repositoryIds: string[]
-) {
-  if (repositoryIds.length === 0) {
-    return triggers;
-  }
-
-  const repositoryIdSet = new Set(repositoryIds);
-
-  return triggers.filter((trigger) => {
-    const parsed = scheduleTargetsRepositoryIdsSchema.safeParse(
-      trigger.targets
-    );
-
-    if (!parsed.success) {
-      return false;
-    }
-
-    return parsed.data.repositoryIds.some((repositoryId) =>
-      repositoryIdSet.has(repositoryId)
-    );
-  });
-}
 
 schedulesRoutes.get("/:organizationId/schedules", async (c) => {
   const orgId = getOrganizationId(c);
@@ -458,65 +285,18 @@ schedulesRoutes.openapi(getSchedulesRoute, async (c) => {
   }
 
   const { repositoryIds } = c.req.valid("query");
-  const triggers = await db.query.contentTriggers.findMany({
-    where: and(
-      eq(contentTriggers.organizationId, orgId),
-      eq(contentTriggers.sourceType, "cron")
-    ),
-    orderBy: [desc(contentTriggers.createdAt)],
-  });
-
-  const triggerIds = triggers.map((trigger) => trigger.id);
-  const lookbackWindows =
-    triggerIds.length > 0
-      ? await db.query.contentTriggerLookbackWindows.findMany({
-          where: inArray(contentTriggerLookbackWindows.triggerId, triggerIds),
-        })
-      : [];
-
-  const lookbackWindowByTriggerId = new Map(
-    lookbackWindows.map((item) => [item.triggerId, item.window])
-  );
-  const filteredTriggers = filterByRepositoryIds(triggers, repositoryIds);
-
-  const schedules = filteredTriggers
-    .map((trigger) =>
-      safeSerializeSchedule({
-        ...trigger,
-        lookbackWindow:
-          lookbackWindowByTriggerId.get(trigger.id) ?? "last_7_days",
-      })
-    )
-    .filter((schedule) => schedule !== null);
-
-  const allRepositoryIds = [
-    ...new Set(schedules.flatMap((schedule) => schedule.targets.repositoryIds)),
-  ];
-  const repositories =
-    allRepositoryIds.length > 0
-      ? await db
-          .select({
-            id: githubIntegrations.id,
-            owner: githubIntegrations.owner,
-            repo: githubIntegrations.repo,
-            defaultBranch: githubIntegrations.defaultBranch,
-          })
-          .from(githubIntegrations)
-          .where(inArray(githubIntegrations.id, allRepositoryIds))
-      : [];
-
-  const repositoryMap = Object.fromEntries(
-    repositories
-      .filter((repository) => repository.owner && repository.repo)
-      .map((repository) => [
-        repository.id,
-        repository.defaultBranch?.trim()
-          ? `${repository.owner}/${repository.repo} · ${repository.defaultBranch.trim()}`
-          : `${repository.owner}/${repository.repo}`,
-      ])
+  const result = await runScheduleProgram(
+    listSchedules({ db, organizationId: orgId, repositoryIds })
   );
 
-  return c.json({ schedules, repositoryMap, organization }, 200);
+  if (result._tag === "Failure") {
+    throw result.failure;
+  }
+
+  return c.json(
+    { ...result.success, organization },
+    200
+  );
 });
 
 schedulesRoutes.openapi(createScheduleRoute, async (c) => {
@@ -657,185 +437,40 @@ schedulesRoutes.openapi(patchScheduleRoute, async (c) => {
   }
 
   const { scheduleId } = c.req.valid("param");
-  const input = c.req.valid("json");
+  const body = c.req.valid("json");
   const env = (c.env ?? {}) as {
     QSTASH_TOKEN?: string;
     WORKFLOW_BASE_URL?: string;
   };
-  const normalized = normalizeSchedule(input);
-  const dedupeHash = hashSchedule(input);
-  const duplicate = await db.query.contentTriggers.findFirst({
-    where: and(
-      eq(contentTriggers.organizationId, orgId),
-      eq(contentTriggers.dedupeHash, dedupeHash),
-      ne(contentTriggers.id, scheduleId)
-    ),
-  });
 
-  if (duplicate) {
-    return c.json({ error: "Duplicate schedule" }, 409);
-  }
-
-  if (input.enabled) {
-    const missingTargets = await ensureScheduleTargetsExist(
+  const result = await runScheduleProgram(
+    patchSchedule({
       db,
-      orgId,
-      normalized.targets.repositoryIds,
-      "Cannot enable schedule: one or more integrations have been deleted"
-    );
+      organizationId: orgId,
+      scheduleId,
+      body,
+      env,
+    })
+  );
 
-    if (missingTargets) {
-      return c.json({ error: missingTargets.error }, 400);
+  if (result._tag === "Failure") {
+    const failure = result.failure;
+    if (failure._tag === "ScheduleNotFoundError") {
+      return c.json({ error: "Schedule not found" }, 404);
     }
+    if (failure._tag === "ScheduleDuplicateError") {
+      return c.json({ error: "Duplicate schedule" }, 409);
+    }
+    if (failure._tag === "ScheduleMissingTargetsError") {
+      return c.json({ error: failure.message }, 400);
+    }
+    if (failure._tag === "ScheduleQstashError") {
+      return c.json({ error: failure.message }, failure.status);
+    }
+    throw failure;
   }
 
-  const affectedQstashIds = new Set<string>();
-  let committed = false;
-  try {
-    const response = await db.transaction(async (tx) => {
-      // Lock before reading or changing remote state. Concurrent PATCHes must
-      // observe the preceding request's committed configuration.
-      const [existing] = await tx
-        .select()
-        .from(contentTriggers)
-        .where(
-          and(
-            eq(contentTriggers.id, scheduleId),
-            eq(contentTriggers.organizationId, orgId),
-            eq(contentTriggers.sourceType, "cron")
-          )
-        )
-        .for("update");
-
-      if (!existing) {
-        return c.json({ error: "Schedule not found" }, 404);
-      }
-
-      let qstashScheduleId: string | null = null;
-      if (input.enabled) {
-        // Choose the ID before sending so even a lost response is recoverable.
-        qstashScheduleId = existing.qstashScheduleId ?? crypto.randomUUID();
-        affectedQstashIds.add(qstashScheduleId);
-        const returnedId = await createQstashSchedule(env, {
-          triggerId: scheduleId,
-          cron: buildCronExpression(normalized.sourceConfig.cron),
-          scheduleId: qstashScheduleId,
-        });
-        affectedQstashIds.add(returnedId);
-        if (returnedId !== qstashScheduleId) {
-          throw new Error("QStash returned an unexpected schedule ID");
-        }
-      } else if (existing.qstashScheduleId) {
-        affectedQstashIds.add(existing.qstashScheduleId);
-        await deleteQstashScheduleWithRetry(env, existing.qstashScheduleId);
-      }
-
-      const [updatedTrigger] = await tx
-        .update(contentTriggers)
-        .set({
-          name: input.name.trim() || existing.name || DEFAULT_SCHEDULE_NAME,
-          sourceType: "cron",
-          sourceConfig: normalized.sourceConfig,
-          targets: normalized.targets,
-          outputType: input.outputType,
-          outputConfig: input.outputConfig ?? null,
-          dedupeHash,
-          enabled: input.enabled,
-          autoPublish: input.autoPublish,
-          qstashScheduleId,
-          updatedAt: new Date(),
-        })
-        .where(
-          and(
-            eq(contentTriggers.id, scheduleId),
-            eq(contentTriggers.organizationId, orgId)
-          )
-        )
-        .returning();
-
-      if (!updatedTrigger) {
-        throw new Error("Failed to update schedule");
-      }
-
-      await tx
-        .insert(contentTriggerLookbackWindows)
-        .values({
-          triggerId: scheduleId,
-          window: input.lookbackWindow,
-        })
-        .onConflictDoUpdate({
-          target: contentTriggerLookbackWindows.triggerId,
-          set: {
-            window: input.lookbackWindow,
-            updatedAt: new Date(),
-          },
-        });
-
-      const schedule = serializeSchedule({
-        ...updatedTrigger,
-        lookbackWindow: input.lookbackWindow,
-      });
-      return c.json({ schedule, organization }, 200);
-    });
-    committed = true;
-    return response;
-  } catch (error) {
-    if (!committed && affectedQstashIds.size > 0) {
-      try {
-        await db.transaction(async (tx) => {
-          // The failed transaction released its lock. Re-read under a new lock:
-          // another request may have committed while recovery was waiting.
-          const [current] = await tx
-            .select()
-            .from(contentTriggers)
-            .where(
-              and(
-                eq(contentTriggers.id, scheduleId),
-                eq(contentTriggers.organizationId, orgId),
-                eq(contentTriggers.sourceType, "cron")
-              )
-            )
-            .for("update");
-          const currentQstashId = current?.enabled
-            ? current.qstashScheduleId
-            : null;
-
-          if (currentQstashId && current) {
-            const config = scheduleSourceConfigSchema.parse(
-              current.sourceConfig
-            );
-            const restoredId = await createQstashSchedule(env, {
-              triggerId: scheduleId,
-              cron: buildCronExpression(config.cron),
-              scheduleId: currentQstashId,
-            });
-            if (restoredId !== currentQstashId) {
-              await deleteQstashScheduleWithRetry(env, restoredId);
-              throw new Error("QStash returned an unexpected restoration ID");
-            }
-          }
-
-          for (const affectedId of affectedQstashIds) {
-            if (affectedId !== currentQstashId) {
-              await deleteQstashScheduleWithRetry(env, affectedId);
-            }
-          }
-        });
-      } catch (recoveryError) {
-        logError(
-          `Failed to reconcile QStash schedules ${[...affectedQstashIds].join(", ")} for schedule ${scheduleId}; retry PATCH to heal`,
-          recoveryError
-        );
-      }
-    }
-
-    logError("Failed to update schedule", error);
-    const mapped = mapQstashError(error);
-    if (mapped.status === 400) {
-      return c.json({ error: mapped.error }, 400);
-    }
-    return c.json({ error: "Failed to update schedule" }, 500);
-  }
+  return c.json({ schedule: result.success, organization }, 200);
 });
 
 schedulesRoutes.openapi(deleteScheduleRoute, async (c) => {
